requirements_dev.txt
pip==19.2.3
bump2version==0.5.11
wheel==0.33.6
watchdog==0.9.0
flake8==3.7.8
tox==3.14.0
coverage==4.5.4
Sphinx==1.8.5
twine==1.14.0

spacy>=2.3.2
streamlit>=0.62.1
torchserve>=0.2.0
torch-model-archiver>=0.2.0
transformers>=3.1.0

